64-bit-friendlify.

+++ tamarin-central/core/avmplusDebugger.cpp
@@ -378,8 +378,8 @@ namespace avmplus
 		scanResources(abc, pool);
 
 		// build a bridging table from pools to abcs
-		int index = abcList.size();
-		pool2abcIndex.add(pool, (const void*)index);
+        uintptr index = abcList.size();
+		pool2abcIndex.add(pool, (const void *)index);
 
 		// at this point our abc object has been populated with
 		// source file objects and should ready to go.
